Matechi-master/requirements.txt
adminlte-base==0.2.0
arabic-reshaper==2.1.3
arrow==1.0.2
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-adminlte-ui==1.5.1
django-baton-v2==1.5.10
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-online-users==0.3
django-static-jquery-ui==1.12.1.1
django-tabbed-changeform-admin==0.1.6
django-treebeard==4.3
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
MarkupSafe==1.1.1
oauth2client==1.5.2
packaging==20.9
Pillow==8.1.0
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.66
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
